Move main behavior to root command internals
Future-proofs any further commands being added by coupling the main
behavior with the CLI behavior, meaning it will only run when the root
command is run.

For #41
Fixes #17
